pyproject.toml
[project]
name = "attendance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.24",
]
